Move scenario sources to sub directory
