+++ src/network/network.cpp
@@ -66,7 +66,7 @@ extern "C" {
 #include "../localisation/localisation.h"
 #include "../management/finance.h"
 #include "../network/http.h"
-#include "../scenario.h"
+#include "../scenario/scenario.h"
 #include "../windows/error.h"
 #include "../util/util.h"
 #include "../cheats.h"
