+++ src/object/StexObject.h
@@ -20,7 +20,7 @@
 
 extern "C"
 {
-    #include "../scenario.h"
+    #include "../scenario/scenario.h"
 }
 
 class StexObject final : public Object
