+++ src/rct2/S6Exporter.cpp
@@ -39,7 +39,7 @@ extern "C"
     #include "../rct2.h"
     #include "../ride/ride.h"
     #include "../ride/ride_ratings.h"
-    #include "../scenario.h"
+    #include "../scenario/scenario.h"
     #include "../util/sawyercoding.h"
     #include "../util/util.h"
     #include "../world/climate.h"
