+++ src/rct2/S6Exporter.h
@@ -23,7 +23,7 @@
 
 extern "C"
 {
-    #include "../scenario.h"
+    #include "../scenario/scenario.h"
     #include "../object_list.h"
 }
