+++ src/rct2/S6Importer.cpp
@@ -35,7 +35,7 @@ extern "C"
     #include "../rct2.h"
     #include "../ride/ride.h"
     #include "../ride/ride_ratings.h"
-    #include "../scenario.h"
+    #include "../scenario/scenario.h"
     #include "../util/sawyercoding.h"
     #include "../world/climate.h"
     #include "../world/map_animation.h"
