+++ src/ride/ride.c
@@ -34,7 +34,7 @@
 #include "../peep/peep.h"
 #include "../peep/staff.h"
 #include "../rct1.h"
-#include "../scenario.h"
+#include "../scenario/scenario.h"
 #include "../util/util.h"
 #include "../windows/error.h"
 #include "../world/banner.h"
