+++ src/scenario/ScenarioSources.h
@@ -16,7 +16,7 @@
 
 #pragma once
 
-#include "common.h"
+#include "../common.h"
 
 typedef struct source_desc
 {
