+++ src/scenario/scenario.c
@@ -14,37 +14,37 @@
  *****************************************************************************/
 #pragma endregion
 
-#include "audio/audio.h"
-#include "cheats.h"
-#include "config.h"
-#include "game.h"
-#include "interface/viewport.h"
-#include "localisation/date.h"
-#include "localisation/localisation.h"
-#include "management/award.h"
-#include "management/finance.h"
-#include "management/marketing.h"
-#include "management/news_item.h"
-#include "management/research.h"
-#include "network/network.h"
-#include "object.h"
-#include "object_list.h"
-#include "OpenRCT2.h"
-#include "peep/staff.h"
-#include "platform/platform.h"
-#include "ride/ride.h"
+#include "../audio/audio.h"
+#include "../cheats.h"
+#include "../config.h"
+#include "../game.h"
+#include "../interface/viewport.h"
+#include "../localisation/date.h"
+#include "../localisation/localisation.h"
+#include "../management/award.h"
+#include "../management/finance.h"
+#include "../management/marketing.h"
+#include "../management/news_item.h"
+#include "../management/research.h"
+#include "../network/network.h"
+#include "../object.h"
+#include "../object_list.h"
+#include "../OpenRCT2.h"
+#include "../peep/staff.h"
+#include "../platform/platform.h"
+#include "../rct1.h"
+#include "../ride/ride.h"
+#include "../util/sawyercoding.h"
+#include "../util/util.h"
+#include "../world/climate.h"
+#include "../world/map.h"
+#include "../world/park.h"
+#include "../world/scenery.h"
+#include "../world/sprite.h"
+#include "../world/water.h"
 #include "scenario.h"
 #include "ScenarioRepository.h"
 #include "ScenarioSources.h"
-#include "util/sawyercoding.h"
-#include "util/util.h"
-#include "world/climate.h"
-#include "world/map.h"
-#include "world/park.h"
-#include "world/scenery.h"
-#include "world/sprite.h"
-#include "world/water.h"
-#include "rct1.h"
 
 const rct_string_id ScenarioCategoryStringIds[SCENARIO_CATEGORY_COUNT] = {
 	STR_BEGINNER_PARKS,
