+++ src/title/TitleSequencePlayer.cpp
@@ -21,8 +21,8 @@
 #include "../core/Math.hpp"
 #include "../core/Path.hpp"
 #include "../core/String.hpp"
-#include "../ScenarioRepository.h"
-#include "../ScenarioSources.h"
+#include "../scenario/ScenarioRepository.h"
+#include "../scenario/ScenarioSources.h"
 #include "TitleSequence.h"
 #include "TitleSequenceManager.h"
 #include "TitleSequencePlayer.h"
