+++ src/windows/dropdown.c
@@ -19,7 +19,7 @@
 #include "../interface/window.h"
 #include "../localisation/localisation.h"
 #include "../rct2.h"
-#include "../scenario.h"
+#include "../scenario/scenario.h"
 #include "../sprites.h"
 #include "dropdown.h"
