+++ src/windows/editor_object_selection.c
@@ -33,7 +33,7 @@
 #include "../ride/track.h"
 #include "../ride/track_data.h"
 #include "../ride/track_design.h"
-#include "../scenario.h"
+#include "../scenario/scenario.h"
 #include "../util/util.h"
 #include "../world/footpath.h"
 #include "dropdown.h"
