+++ src/windows/loadsave.c
@@ -23,7 +23,7 @@
 #include "../interface/window.h"
 #include "../localisation/localisation.h"
 #include "../network/network.h"
-#include "../scenario.h"
+#include "../scenario/scenario.h"
 #include "../title/TitleScreen.h"
 #include "../util/util.h"
 #include "../windows/error.h"
