+++ src/windows/title_editor.c
@@ -25,8 +25,8 @@
 #include "../localisation/localisation.h"
 #include "../peep/peep.h"
 #include "../peep/staff.h"
-#include "../scenario.h"
-#include "../ScenarioSources.h"
+#include "../scenario/scenario.h"
+#include "../scenario/ScenarioSources.h"
 #include "../sprites.h"
 #include "../title/TitleScreen.h"
 #include "../title/TitleSequence.h"
