+++ src/windows/title_scenarioselect.c
@@ -18,8 +18,8 @@
 #include "../audio/audio.h"
 #include "../localisation/date.h"
 #include "../localisation/localisation.h"
-#include "../ScenarioRepository.h"
-#include "../ScenarioSources.h"
+#include "../scenario/ScenarioRepository.h"
+#include "../scenario/ScenarioSources.h"
 #include "../sprites.h"
 #include "../interface/widget.h"
 #include "../interface/window.h"
