+++ src/world/balloon.c
@@ -16,7 +16,7 @@
 
 #include "../audio/audio.h"
 #include "../network/network.h"
-#include "../scenario.h"
+#include "../scenario/scenario.h"
 #include "../util/util.h"
 #include "sprite.h"
