+++ src/world/climate.c
@@ -23,7 +23,7 @@
 #include "../interface/window.h"
 #include "../localisation/date.h"
 #include "../rct2.h"
-#include "../scenario.h"
+#include "../scenario/scenario.h"
 #include "../sprites.h"
 #include "../util/util.h"
 #include "climate.h"
