+++ src/world/particle.c
@@ -16,7 +16,7 @@
 
 #include "../audio/audio.h"
 #include "../util/util.h"
-#include "../scenario.h"
+#include "../scenario/scenario.h"
 #include "sprite.h"
 
 /**
