+++ src/world/sprite.c
@@ -22,7 +22,7 @@
 #include "../localisation/date.h"
 #include "../localisation/localisation.h"
 #include "../OpenRCT2.h"
-#include "../scenario.h"
+#include "../scenario/scenario.h"
 #include "fountain.h"
 #include "sprite.h"
